Leave the moved-out container in a well-defined state
Clang's static analyzer warns that the destructor should not
be called on the container that's had its content moved out.

This appears to be a valid issue, given the container should
be left in a well-defined state. For example, the container
could just as easily have more content moved back into it;
or it could allocate new content.

Calling the destructor on the moved-out container does not
leave it in a workable state.

This logic is also confirmed in the last paragraph here:
https://www.internalpointers.com/post/c-rvalue-references-and-move-semantics-beginners

---

A typical move constructor:

Holder(Holder&& other) // <-- rvalue reference in input
{
    m_data = other.m_data;   // (1)
    m_size = other.m_size;
    other.m_data = nullptr;  // (2)
    other.m_size = 0;
}

It takes in input an rvalue reference to another Holder
object. This is the key part: being an rvalue reference, we
can modify it

So let's steal its data first (1), then set it to null (2).
No deep copies here, we have just moved resources around!

It's important to set the rvalue reference data to some
valid state (2) to prevent it from being accidentally
deleted when the temporary object dies: our Holder
destructor calls delete[] m_data, remember?

In general, for reasons that will become more clear in a few
paragraphs, it's a good idea to always leave the objects
being stolen from in some well-defined state.

---